修复了在高android版本中无法删除歌曲的bug。 修复了MIUI设备上媒体通知异常的bug。

+++ app/src/main/java/com/zxhhyj/music/ui/dialog/DeleteSongDialog.kt
@@ -1,15 +1,25 @@
 package com.zxhhyj.music.ui.dialog
 
+import android.content.ContentUris
+import android.os.Build
+import android.provider.MediaStore
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.IntentSenderRequest
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.clickable
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalLifecycleOwner
 import androidx.compose.ui.res.stringResource
+import androidx.lifecycle.lifecycleScope
+import com.zxhhyj.music.MainApplication
 import com.zxhhyj.music.R
 import com.zxhhyj.music.logic.bean.SongBean
 import com.zxhhyj.music.logic.repository.AndroidMediaLibRepository
 import com.zxhhyj.music.logic.repository.WebDavMediaLibRepository
 import com.zxhhyj.ui.view.YesNoDialog
+import kotlinx.coroutines.launch
 import java.io.File
 
 @Composable
@@ -17,6 +27,16 @@ fun DeleteSongDialog(
     onDismissRequest: () -> Unit,
     songBean: SongBean
 ) {
+    val deleteLauncher = rememberLauncherForActivityResult(
+        contract = ActivityResultContracts.StartIntentSenderForResult(),
+        onResult = {
+            if (it.resultCode == -1) {
+                AndroidMediaLibRepository.removeSong(songBean as SongBean.Local)
+                onDismissRequest.invoke()
+            }
+        }
+    )
+    val lifecycleCoroutineScope = LocalLifecycleOwner.current.lifecycleScope
     YesNoDialog(
         onDismissRequest = onDismissRequest,
         title = stringResource(id = R.string.delete),
@@ -26,20 +46,40 @@ fun DeleteSongDialog(
                 modifier = Modifier.clickable {
                     when (songBean) {
                         is SongBean.Local -> {
-                            val file = File(songBean.data)
-                            file.delete()
-                            AndroidMediaLibRepository.removeSong(songBean)
+                            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+                                lifecycleCoroutineScope.launch {
+                                    val deleteUri =
+                                        ContentUris.withAppendedId(
+                                            MediaStore.Audio.Media.EXTERNAL_CONTENT_URI,
+                                            songBean.id
+                                        )
+                                    val pendingIntent = MediaStore.createDeleteRequest(
+                                        MainApplication.context.contentResolver,
+                                        listOf(deleteUri)
+                                    )
+                                    val request =
+                                        IntentSenderRequest.Builder(pendingIntent.intentSender)
+                                            .build()
+                                    deleteLauncher.launch(request)
+                                }
+                            } else {
+                                val file = File(songBean.data)
+                                if (file.delete()) {
+                                    AndroidMediaLibRepository.removeSong(songBean)
+                                }
+                            }
                         }
 
                         is SongBean.WebDav -> {
                             val file = File(songBean.data)
-                            file.delete()
-                            WebDavMediaLibRepository.removeSong(songBean)
+                            if (file.delete()) {
+                                WebDavMediaLibRepository.removeSong(songBean)
+                                onDismissRequest.invoke()
+                            }
                         }
 
                         else -> {}
                     }
-                    onDismissRequest.invoke()
                 })
             if (songBean is SongBean.Local) {
                 Text(
